backend/requirements.txt
# fastapi
# uvicorn
# motor
# passlib[bcrypt]
# python-jose
# pydantic
# requests
# fpdf
# PyJWT
# python-dateutil


# fastapi
# uvicorn[standard]
# motor
# passlib[bcrypt]
# python-jose[cryptography]
# pydantic[email]
# requests
# fpdf
# python-dateutil
# python-dotenv


fastapi
uvicorn[standard]
motor
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
fpdf
python-dateutil
python-dotenv
google-generativeai